Version 4: soporte para coeficientes decimales usando double

+++ src/ecuacionsegundogrado/Main.java
@@ -7,28 +7,25 @@ public class Main {
 
         Scanner sc = new Scanner(System.in);
 
-        int a, b, c;
+        double a, b, c;
         double discriminante, x1, x2;
 
         System.out.print("Ingrese a: ");
-        a = sc.nextInt();
+        a = sc.nextDouble();
 
         System.out.print("Ingrese b: ");
-        b = sc.nextInt();
+        b = sc.nextDouble();
 
         System.out.print("Ingrese c: ");
-        c = sc.nextInt();
+        c = sc.nextDouble();
 
-        // VALIDACIÓN 1: a no puede ser 0
         if (a == 0) {
             System.out.println("No es una ecuacion de segundo grado.");
             return;
         }
 
-        // CÁLCULO DEL DISCRIMINANTE
         discriminante = (b * b) - (4 * a * c);
 
-        // CASOS
         if (discriminante > 0) {
 
             x1 = (-b + Math.sqrt(discriminante)) / (2.0 * a);
